- Update ThanhToan, ChiTietPhong

+++ DTO/DichVuPhongDTO.cs
@@ -15,10 +15,10 @@ namespace DTO
 
         public DateTime ThoiGian { get; set; }
 
-        public int SoLuong { get; set; }
+        public double SoLuong { get; set; }
         public double DonGia { get; set; }
 
-        public DichVuPhongDTO(int maDVP, int maThuePhong,int maDV, DateTime thoiGian, int soLuong, double donGia)
+        public DichVuPhongDTO(int maDVP, int maThuePhong,int maDV, DateTime thoiGian, double soLuong, double donGia)
         {
             MaDVP = maDVP;
             MaThuePhong = maThuePhong;
